Add files via upload
Luhn's algorithm calculates, from a number (or a sequence of digits), a check key (called checksum), this key is a digit which is dependent on the others.
What is the Luhn Algorithm for?
Luhn makes it possible to check numbers (credit card, SIRET, etc.) thanks to its control key (a digit which makes it possible to check the others digits). If a character is misread or badly written, then Luhn's algorithm will detect this error.
Luhn is known because MasterCard, American Express (AMEX), Visa and all credit cards use it.